refactor: フェーズ3 全ゲーム (7-11) + フェーズ4 統合リファクタリング
- Air Hockey: hooks/components 分離、AirHockeyPage を薄いラッパーに
- IPNE: presentation 層分割 (screens/hooks/config)、IpnePage を薄いラッパーに
- Risk LCD: フェーズ別フック分割 (phases/)、useGameEngine をオーケストレーターに
- Non-Brake Descent: renderers をサブディレクトリ化 (environment/entities/effects/ui)
- Agile Quiz Sugoroku: styles.ts (1,070行) を5モジュールに分割 + コンポーネントテスト追加
- フェーズ4: ローカル math 関数の共通化、全テスト通過・ビルド確認
