+++ src/features/air-hockey/index.ts
@@ -1,0 +1,8 @@
+export { default as AirHockeyGame } from './AirHockeyGame';
+export { Renderer } from './renderer';
+export { TitleScreen } from './components/TitleScreen';
+export { Scoreboard } from './components/Scoreboard';
+export { Field } from './components/Field';
+export { ResultScreen } from './components/ResultScreen';
+export { useInput } from './hooks/useInput';
+export { useGameLoop } from './hooks/useGameLoop';
