+++ src/features/non-brake-descent/renderers/index.tsx
@@ -1,0 +1,11 @@
+// 環境描画
+export { CloudRenderer, BuildingRenderer } from './environment';
+
+// エンティティ描画
+export { ObstacleRenderer, RampRenderer, PlayerRenderer } from './entities';
+
+// エフェクト描画
+export { ParticlesRenderer, ScorePopupsRenderer, NearMissRenderer, DangerVignette } from './effects';
+
+// UI描画
+export { UIOverlay, CountdownOverlay, ScreenOverlay, TouchButton, MobileControls } from './ui';
